Logistic Regression
Implemented logistic regression for the following scenarios:

  1. Linear decision boundary to predict a student's admission to a university based on performance on 2 tests
  2. Non-linear decision boundary for microchips based on 2 performance metrics

For both problems, data was first visualized with scatter plots.  A sigmoid function and the logistic cost function were implemented.  Cost function optimization is handled by the octave function "fminunc" to find the minimum of an unconstrained function.  For the non-linear decision boundary the function "mapFeature.m" was written to create 28 polynomial features to fit the non-linear decision boundary.